--- js/src/components/cyo-incentive-picker/cyoi-radio-control.js
@@ -1,43 +1,0 @@
-/**
- * External dependencies
- */
-import classnames from 'classnames';
-import { RadioControl } from '@wordpress/components';
-
-/**
- * Internal dependencies
- */
-import useAdsCurrency from '~/hooks/useAdsCurrency';
-import styles from './cyo-incentive-picker.module.scss';
-
-const CYOIRadioControl = ( {
-	label,
-	value,
-	requirement,
-	selected,
-	className,
-	...rest
-} ) => {
-	const { formatAmount } = useAdsCurrency();
-
-	return (
-		// eslint-disable-next-line jsx-a11y/label-has-associated-control
-		<label className={ styles.radioGroup }>
-			<RadioControl
-				{ ...rest }
-				label={ formatAmount( label ) }
-				className={ classnames( className ) }
-				checked={ selected === value }
-				options={ [ { value } ] }
-				hideLabelFromVision
-				help=""
-				aria-label={ value }
-			/>
-			<span className={ classnames( styles.option ) }>
-				{ formatAmount( label ) }
-			</span>
-		</label>
-	);
-};
-
-export default CYOIRadioControl;
