--- js/src/components/cyo-incentive-picker/index.scss
@@ -1,7 +1,0 @@
-.gla-cyoi-section {
-
-    .gla-subsection-title {
-        font-weight: 400;
-        margin-bottom: 4px;
-    }
-}
